trigger mem leaks
Ignore-this: 47e5be238f657a747884ee8387d55c56

darcs-hash:20100128130716-68901-45546b2c84922cbaa429715a55e2714b34498819.gz

+++ kernel/tr/triggers/triggers.cpp
@@ -692,7 +692,7 @@ trigger_cell_xptr create_trigger (
     t_scmnodes sobj = execute_abs_path_expr(schemaroot, trigger_path, NULL, NULL);
     //III. For each schema node found (sn_obj)
     std::vector<xptr> start_nodes;
-    for (int i = 0; i < sobj.size(); i++)
+    for (size_t i = 0; i < sobj.size(); i++)
     {
         sobj[i].modify()->trigger_list.add(trc.ptr());
         RECOVERY_CRASH;
